requirements.txt
opencv-python
numpy
moviepy
psycopg2-binary
boto3
python-dotenv